Develop

+++ ObliGaitanBordaAnon/ObliGaitanBordaAnon/Models/Cotizacione.cs
@@ -9,7 +9,7 @@ public partial class Cotizacione
 
     public string? NombreDivisa { get; set; }
 
-    public int? CotizacionDivisa { get; set; }
+    public double? CotizacionDivisa { get; set; }
 
     public virtual ICollection<Pago> Pagos { get; set; } = new List<Pago>();
 }
